+++ src/main/java/insane96mcp/survivalreimagined/data/lootmodifier/SRGlobalLootModifierProvider.java
@@ -1,7 +1,7 @@
 package insane96mcp.survivalreimagined.data.lootmodifier;
 
 import insane96mcp.survivalreimagined.module.farming.feature.Crops;
-import insane96mcp.survivalreimagined.module.items.feature.FlintTools;
+import insane96mcp.survivalreimagined.module.items.feature.FlintExpansion;
 import insane96mcp.survivalreimagined.module.mobs.feature.Equipment;
 import insane96mcp.survivalreimagined.module.world.feature.Loot;
 import net.minecraft.data.PackOutput;
@@ -17,6 +17,6 @@ public class SRGlobalLootModifierProvider extends GlobalLootModifierProvider {
         Crops.addGlobalLoot(this);
         Equipment.addGlobalLoot(this);
         Loot.addGlobalLoot(this);
-        FlintTools.addGlobalLoot(this);
+        FlintExpansion.addGlobalLoot(this);
     }
 }