+++ src/main/java/insane96mcp/survivalreimagined/setup/SRBlocks.java
@@ -7,4 +7,7 @@ import net.minecraftforge.registries.ForgeRegistries;
 
 public class SRBlocks {
     public static final DeferredRegister<Block> REGISTRY = DeferredRegister.create(ForgeRegistries.BLOCKS, SurvivalReimagined.MOD_ID);
+
+    /*
+    public static final BlockWithItem POLISHED_FLINT_BLOCK = new BlockWithItem("polished_flint_block", new Block(BlockBehaviour.Properties.of(Material.STONE, MaterialColor.STONE).requiresCorrectToolForDrops().strength(1.5F, 5.0F)));*/
 }