+++ changelog.md
@@ -3,6 +3,7 @@
 ## Upcoming
 * Added Anvil feature
   * An anvil dropping on a block may transform it
+* Added Flint Block and Polished Flint Block
 * Fishing changes
   * You can no longer get fishes in small bodies of water
   * Junk chance is greatly increased